requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
html2text
python-dotenv
google-generativeai
beautifulsoup4
schedule